Nuking the linked list
For reasons which I will never know the #defined linked list didnt't work with my code. No idea why but I need to get a move on with this so I can actually begin writing some Vulkan code. Instead I'm going to try and brute force it and not try to do it in a clever way.